Import and pull projects from Overleaf.
Add the option to import projects from Overleaf.
The password is stored in shared memory and never hits the disk.
When the clone has been successful, it is stored there until the container is stopped.
It is also possible to update (pull) documents later, when the files have been imported as projects.