Status ok_exitcrash
On alpine linux 3.16, some conversions crash non-deterministically.
The crash happens at the end, while memory is freed. The destination
file has been completely created though, but the
term "conversion complete" has not been written out.
Detect these cases and and make sure that such conversions are detected.
If the file is complete, the conversion is not interpreded as fatal error, but
rather the status 'ok_exitcrash' is given.